LIBITD-2455. Added ".dockerignore" file
Added ".dockerignore" file to allow "Dockerfile" to simply copy
everything in the directory (except files in the ".dockerignore" file
into the Docker image.

This seemed easier and less error-prone than specifying the individual
files/directories to add in the COPY command "Dockerfile", and
more consistent with how other Dockerfiles are constructed.

Also updated the ".gitignore" file to ignore the ".ruff_cache" directory
(it was basically being ignored already, because of a ".gitignore"
file in the subdirectory, but this dims the ".ruff_cache" directory
itself in VS Code).

https://umd-dit.atlassian.net/browse/LIBITD-2455