LIBITD-2455. Remove indent for JSON log formatter
Having an "indent" field on the JSON log formatter is causing each field
in the JSON log message to appear on a separate line, and Splunk is
treating each line as a separate log entry (so a single "logical" line
of log output is being spread over multiple Splunk log entries).

Removing the "indent" field so JSON is sent as one line to Splunk.

https://umd-dit.atlassian.net/browse/LIBITD-2455