LIBITD-2455. Added Jenkinsfile for continuous integration
Took the "Jenkinsfile" from "umd-lib/catalog-searcher", to enable
continuous integration for this application.

The "ruff" tool only detects some of the warnings/errors that
"pycodestyle" catches, so running both tools to catch the maximum
amount of errors.

Updated the "pycodestyle" invocation to produce a code coverage report.

Added "pycodestyle" to "test" dependencies in "pyproject.toml" so
it could be run in the Jenkinsfile.

https://umd-dit.atlassian.net/browse/LIBITD-2455